ref(onboarding): Split flask onboarding docs
